+++ client/cmd/outview-gui/main.go
@@ -17,7 +17,7 @@ import (
 )
 
 var (
-	Version   = "1.0.0"
+	Version   = "1.0.2-SNAPSHOT"
 	BuildDate = "unknown"
 )